yolov5/requirements.txt
albumentations==1.1.0
coremltools==5.1.0
Flask==2.0.1
matplotlib==3.4.3
numpy==1.21.4
onnx==1.10.2
onnxruntime==1.9.0
opencv_python==4.5.3.56
pafy==0.5.5
pandas==1.3.2
Pillow==8.4.0
protobuf==3.19.1
pycocotools==2.0.2
PyYAML==6.0
requests==2.26.0
scipy==1.7.1
seaborn==0.11.2
setuptools==58.0.4
tensorflow==2.7.0
tensorflowjs==3.11.0
thop==0.0.31.post2005241907
torch==1.8.0+cu111
torchvision==0.9.0+cu111
tqdm==4.62.1
Werkzeug==2.0.1
